+++ packages/widgetbook/test/src/addons/utils/addon_test_helper.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:go_router/go_router.dart';
-import 'package:widgetbook/src/addons/common/common.dart';
+import 'package:widgetbook/src/addons/addons.dart';
 import 'package:widgetbook/src/state/state.dart';
 
 import 'extensions/widget_tester_extension.dart';
@@ -14,44 +13,22 @@ Future<void> testAddon({
 }) async {
   const key = ValueKey('RandomKey');
 
-  final router = GoRouter(
-    routes: [
-      GoRoute(
-        path: '/',
-        name: '/',
-        builder: (context, state) {
-          return WidgetbookScope(
-            state: WidgetbookState(
-              path: state.queryParams['path'] ?? '',
-              previewMode: false,
-              queryParams: {...state.queryParams}, // Copy from UnmodifiableMap
-              addons: [addon],
-              appBuilder: materialAppBuilder,
-              catalog: WidgetbookCatalog.fromDirectories([]),
-            ),
-            child: Scaffold(
-              // Simulates the approximate width of the panel.
-              // It's also required to make some tests work
-              // (not totally clear why).
-              body: SizedBox(
-                width: 500,
-                child: Builder(
-                  key: key,
-                  builder: (context) {
-                    return addon.buildSetting(context);
-                  },
-                ),
-              ),
-            ),
-          );
-        },
-      )
-    ],
-  );
-
   await tester.pumpWidget(
-    MaterialApp.router(
-      routerConfig: router,
+    MaterialApp(
+      home: WidgetbookScope(
+        state: WidgetbookState(
+          queryParams: {},
+          addons: [addon],
+          appBuilder: materialAppBuilder,
+          catalog: WidgetbookCatalog.fromDirectories([]),
+        ),
+        child: Scaffold(
+          body: Builder(
+            key: key,
+            builder: addon.buildSetting,
+          ),
+        ),
+      ),
     ),
   );